% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{mhs_cols}
\alias{mhs_cols}
\title{Extract MHS Colors as Hex Codes}
\usage{
mhs_cols(...)
}
\arguments{
\item{...}{Character names of colors}
}
\value{
Named character vector.
}
\description{
Extract MHS Colors as Hex Codes
}
\examples{
mhs_cols("purple")
mhs_cols("blue", "purple")
}
