% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{mhs_pal}
\alias{mhs_pal}
\title{Return function to interpolate an MHS color palette}
\usage{
mhs_pal(palette = "main", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
A function that can be used to create n number of colors (given a
  pallette).
}
\description{
Return function to interpolate an MHS color palette
}
\examples{

mhs_pal("mhm")
mhs_pal("mhm")(10)
}
