% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhscales-package.R
\docType{package}
\name{mhscales}
\alias{mhscales}
\alias{_PACKAGE}
\alias{mhscales-package}
\title{Missouri Historical Society Scales
  colors.}
\description{
Create scales for ggplot that have the Missouri Historical Society
}
\keyword{internal}
