% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_mhs}
\alias{scale_color_mhs}
\title{Color scale constructor for MHS colors}
\usage{
scale_color_mhs(palette = "mhs", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
A scale_color_*() to be used with ggplot.
}
\description{
Color scale constructor for MHS colors
}
\examples{
library(ggplot2)
library(mhscales)
# Color by discrete variable using default palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Species)) +
  geom_point(size = 4) +
  scale_color_mhs(palette = "mhs_mhm")

# Color by numeric variable with cool palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
  geom_point(size = 4, alpha = .6) +
  scale_color_mhs(palette = "mhm", discrete = FALSE)
}
