% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_mhs}
\alias{scale_fill_mhs}
\title{Fill scale constructor for MHS colors}
\usage{
scale_fill_mhs(palette = "mhs", discrete = TRUE, reverse = FALSE,
  ...)
}
\arguments{
\item{palette}{Character name of palette in drsimonj_palettes}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn(), used respectively when discrete is TRUE or FALSE}
}
\value{
A scale_fill_*() to be used with ggplot.
}
\description{
Fill scale constructor for MHS colors
}
\examples{
library(ggplot2)
library(mhscales)
# Fill by discrete variable with different palette + remove legend (guide)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
  geom_bar() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_fill_mhs(palette = "mhm", guide = "none")

# fill by continuous variable
library(forcats)
library(dplyr)

mpg \%>\%
  count(manufacturer) \%>\%
  mutate(manufacturer = forcats::fct_reorder(manufacturer, n)) \%>\%
  ggplot(aes(x = manufacturer, y = n, fill = n)) +
  geom_col() +
  scale_fill_mhs(palette = "mhs_mhm", discrete = FALSE)+
  coord_flip()
}
